refactor(OrderResultContainer): extract render
- 결과를 렌더링하는 책임을 별도 컴포넌트로 분리
